getDevAndPop/getDevAndPop(Dockerfile)/requirements.txt
flask
google-cloud-storage
pandas
playwright